+++ lib/src/command/command.dart
@@ -252,24 +252,41 @@ class Command {
     subCommand = DecodeGifFileCmd(subCommand, path);
   }
 
-  void encodeGif(
-      {int samplingFactor = 10,
-      DitherKernel dither = DitherKernel.floydSteinberg,
-      bool ditherSerpentine = false}) {
+  void encodeGif({
+    int samplingFactor = 10,
+    DitherKernel dither = DitherKernel.floydSteinberg,
+    @Deprecated('Use ditherScanOrder: DitherScanOrder.serpentine instead. '
+        'This parameter will be removed in a future release.')
+    bool ditherSerpentine = false,
+    DitherScanOrder? ditherScanOrder,
+  }) {
     subCommand = EncodeGifCmd(subCommand,
         samplingFactor: samplingFactor,
         dither: dither,
-        ditherSerpentine: ditherSerpentine);
+        ditherScanOrder: ditherScanOrder ??
+            // ignore: deprecated_member_use_from_same_package
+            (ditherSerpentine
+                ? DitherScanOrder.serpentine
+                : DitherScanOrder.raster));
   }
 
-  void encodeGifFile(String path,
-      {int samplingFactor = 10,
-      DitherKernel dither = DitherKernel.floydSteinberg,
-      bool ditherSerpentine = false}) {
+  void encodeGifFile(
+    String path, {
+    int samplingFactor = 10,
+    DitherKernel dither = DitherKernel.floydSteinberg,
+    @Deprecated('Use ditherScanOrder: DitherScanOrder.serpentine instead. '
+        'This parameter will be removed in a future release.')
+    bool ditherSerpentine = false,
+    DitherScanOrder? ditherScanOrder,
+  }) {
     subCommand = EncodeGifFileCmd(subCommand, path,
         samplingFactor: samplingFactor,
         dither: dither,
-        ditherSerpentine: ditherSerpentine);
+        ditherScanOrder: ditherScanOrder ??
+            // ignore: deprecated_member_use_from_same_package
+            (ditherSerpentine
+                ? DitherScanOrder.serpentine
+                : DitherScanOrder.raster));
   }
 
   // Ico
@@ -801,12 +818,20 @@ class Command {
         maskChannel: maskChannel);
   }
 
-  void ditherImage(
-      {Quantizer? quantizer,
-      DitherKernel kernel = DitherKernel.floydSteinberg,
-      bool serpentine = false}) {
+  void ditherImage({
+    Quantizer? quantizer,
+    DitherKernel kernel = DitherKernel.floydSteinberg,
+    @Deprecated('Use scanOrder: DitherScanOrder.serpentine instead. '
+        'This parameter will be removed in a future release.')
+    bool serpentine = false,
+    DitherScanOrder? scanOrder,
+  }) {
     subCommand = DitherImageCmd(subCommand,
-        quantizer: quantizer, kernel: kernel, serpentine: serpentine);
+        quantizer: quantizer,
+        kernel: kernel,
+        scanOrder: scanOrder ??
+            // ignore: deprecated_member_use_from_same_package
+            (serpentine ? DitherScanOrder.serpentine : DitherScanOrder.raster));
   }
 
   void dotScreen(
@@ -946,16 +971,24 @@ class Command {
         size: size, mode: mode, mask: mask, maskChannel: maskChannel);
   }
 
-  void quantize(
-      {int numberOfColors = 256,
-      QuantizeMethod method = QuantizeMethod.neuralNet,
-      DitherKernel dither = DitherKernel.none,
-      bool ditherSerpentine = false}) {
+  void quantize({
+    int numberOfColors = 256,
+    QuantizeMethod method = QuantizeMethod.neuralNet,
+    DitherKernel dither = DitherKernel.none,
+    @Deprecated('Use ditherScanOrder: DitherScanOrder.serpentine instead. '
+        'This parameter will be removed in a future release.')
+    bool ditherSerpentine = false,
+    DitherScanOrder? ditherScanOrder,
+  }) {
     subCommand = QuantizeCmd(subCommand,
         numberOfColors: numberOfColors,
         method: method,
         dither: dither,
-        ditherSerpentine: ditherSerpentine);
+        ditherScanOrder: ditherScanOrder ??
+            // ignore: deprecated_member_use_from_same_package
+            (ditherSerpentine
+                ? DitherScanOrder.serpentine
+                : DitherScanOrder.raster));
   }
 
   void reinhardTonemap(