+++ lib/src/formats/formats.dart
@@ -495,32 +495,57 @@ Future<Image?> decodeGifFile(String path, {int? frame}) async {
 /// If you know that you have less than 256 colors in your frames
 /// anyway, you should supply a very large [samplingFactor] for maximum
 /// performance.
-Uint8List encodeGif(Image image,
-        {bool singleFrame = false,
-        int repeat = 0,
-        int samplingFactor = 10,
-        DitherKernel dither = DitherKernel.floydSteinberg,
-        bool ditherSerpentine = false}) =>
+///
+/// `ditherSerpentine` is deprecated in favor of [ditherScanOrder] and is
+/// ignored when [ditherScanOrder] is given explicitly.
+Uint8List encodeGif(
+  Image image, {
+  bool singleFrame = false,
+  int repeat = 0,
+  int samplingFactor = 10,
+  DitherKernel dither = DitherKernel.floydSteinberg,
+  @Deprecated('Use ditherScanOrder: DitherScanOrder.serpentine '
+      'instead. This parameter will be removed in a future release.')
+  bool ditherSerpentine = false,
+  DitherScanOrder? ditherScanOrder,
+}) =>
     GifEncoder(
             samplingFactor: samplingFactor,
             dither: dither,
-            ditherSerpentine: ditherSerpentine)
+            ditherScanOrder: ditherScanOrder ??
+                // ignore: deprecated_member_use_from_same_package
+                (ditherSerpentine
+                    ? DitherScanOrder.serpentine
+                    : DitherScanOrder.raster))
         .encode(image, singleFrame: singleFrame);
 
 /// Encode an [image] to a GIF file at the given [path].
-Future<bool> encodeGifFile(String path, Image image,
-    {bool singleFrame = false,
-    int repeat = 0,
-    int samplingFactor = 10,
-    DitherKernel dither = DitherKernel.floydSteinberg,
-    bool ditherSerpentine = false}) async {
+///
+/// `ditherSerpentine` is deprecated in favor of [ditherScanOrder] and is
+/// ignored when [ditherScanOrder] is given explicitly.
+Future<bool> encodeGifFile(
+  String path,
+  Image image, {
+  bool singleFrame = false,
+  int repeat = 0,
+  int samplingFactor = 10,
+  DitherKernel dither = DitherKernel.floydSteinberg,
+  @Deprecated('Use ditherScanOrder: DitherScanOrder.serpentine instead. '
+      'This parameter will be removed in a future release.')
+  bool ditherSerpentine = false,
+  DitherScanOrder? ditherScanOrder,
+}) async {
   if (!supportsFileAccess()) {
     return false;
   }
   final bytes = GifEncoder(
           samplingFactor: samplingFactor,
           dither: dither,
-          ditherSerpentine: ditherSerpentine)
+          ditherScanOrder: ditherScanOrder ??
+              // ignore: deprecated_member_use_from_same_package
+              (ditherSerpentine
+                  ? DitherScanOrder.serpentine
+                  : DitherScanOrder.raster))
       .encode(image, singleFrame: singleFrame);
   return writeFile(path, bytes);
 }